refactor(frontend): Move main services to new folder
