+++ frontend/src/app/components/changelog/changelog.component.ts
@@ -2,7 +2,7 @@ import {Component, OnInit} from '@angular/core';
 import {ActivatedRoute} from '@angular/router';
 import {NgbModal} from '@ng-bootstrap/ng-bootstrap';
 
-import {ChangelogService, Release, REPOS, Repository, Versions} from '../../changelog.service';
+import {ChangelogService, Release, REPOS, Repository, Versions} from '../../services/changelog.service';
 
 @Component({
   selector: 'app-changelog',
