+++ frontend/src/app/components/header/header.component.ts
@@ -2,9 +2,9 @@ import {AfterViewInit, Component, OnInit} from '@angular/core';
 import {ActivatedRoute, Router} from '@angular/router';
 import {KeycloakService} from 'keycloak-angular';
 import {environment} from '../../../environments/environment';
-import {ChangelogService, REPOS, Versions} from '../../changelog.service';
+import {ChangelogService, REPOS, Versions} from '../../services/changelog.service';
 
-import {PrivacyService} from '../../privacy.service';
+import {PrivacyService} from '../../services/privacy.service';
 
 @Component({
   selector: 'app-header',
