+++ frontend/src/app/services/changelog.service.ts
@@ -2,7 +2,7 @@ import {HttpClient} from '@angular/common/http';
 import {Injectable} from '@angular/core';
 import {Observable} from 'rxjs';
 import {map, mapTo, tap} from 'rxjs/operators';
-import {environment} from '../environments/environment';
+import {environment} from '../../environments/environment';
 import {MarkdownService} from './markdown.service';
 
 import {PrivacyService} from './privacy.service';
