+++ frontend/src/app/services/privacy.service.ts
